Moved component styles to own directories to avoid such migration in the future.
