+++ src/button/buttonview.js
@@ -13,7 +13,7 @@ import TooltipView from '../tooltip/tooltipview';
 
 import { getEnvKeystrokeText } from '@ckeditor/ckeditor5-utils/src/keyboard';
 
-import '../../theme/components/button.css';
+import '../../theme/components/button/button.css';
 
 /**
  * The button view class.
