+++ src/dropdown/dropdownview.js
@@ -11,7 +11,7 @@ import View from '../view';
 import FocusTracker from '@ckeditor/ckeditor5-utils/src/focustracker';
 import KeystrokeHandler from '@ckeditor/ckeditor5-utils/src/keystrokehandler';
 
-import '../../theme/components/dropdown.css';
+import '../../theme/components/dropdown/dropdown.css';
 
 /**
  * The dropdown view class.
