+++ src/editorui/editoruiview.js
@@ -12,7 +12,7 @@
 import View from '../view';
 import Template from '../template';
 
-import '../../theme/components/editorui.css';
+import '../../theme/components/editorui/editorui.css';
 
 /**
  * The editor UI view class. Base class for the editor main views.
