+++ src/icon/iconview.js
@@ -11,7 +11,7 @@
 
 import View from '../view';
 
-import '../../theme/components/icon.css';
+import '../../theme/components/icon/icon.css';
 
 /**
  * The icon view class.
