+++ src/list/listview.js
@@ -12,7 +12,7 @@ import FocusTracker from '@ckeditor/ckeditor5-utils/src/focustracker';
 import FocusCycler from '../focuscycler';
 import KeystrokeHandler from '@ckeditor/ckeditor5-utils/src/keystrokehandler';
 
-import '../../theme/components/list.css';
+import '../../theme/components/list/list.css';
 
 /**
  * The list view class.
