+++ theme/components/button/button.css
@@ -3,8 +3,8 @@
  * For licensing, see LICENSE.md.
  */
 
-@import "../mixins/_unselectable.css";
-@import "./tooltip/mixins/_tooltip.css";
+@import "../../mixins/_unselectable.css";
+@import "../tooltip/mixins/_tooltip.css";
 
 .ck-button,
 a.ck-button {
