--- theme/components/editorui.css
@@ -1,4 +1,0 @@
-/*
- * Copyright (c) 2003-2017, CKSource - Frederico Knabben. All rights reserved.
- * For licensing, see LICENSE.md.
- */
