+++ theme/components/label/label.css
